refactor: clean up tool typevars
