--- python/mirascope/llm/tools/async_tool.py
@@ -1,28 +1,0 @@
-"""The `AsyncTool` class for defining async tools that LLMs can request be called."""
-
-from __future__ import annotations
-
-from collections.abc import Awaitable, Callable
-from dataclasses import dataclass
-
-from ..content import ToolCall, ToolOutput
-from ..types import JsonableCovariantT, P
-from .base_tool import BaseTool
-
-
-@dataclass
-class AsyncTool(BaseTool[P, JsonableCovariantT]):
-    """Protocol defining an async tool that can be used by LLMs.
-
-    An AsyncTool represents an async function that can be called by an LLM during a call.
-    It includes metadata like name, description, and parameter schema.
-
-    This class is not instantiated directly but created by the `@tool()` decorator.
-    """
-
-    fn: Callable[P, Awaitable[JsonableCovariantT]]
-    """The async function that implements the tool's functionality."""
-
-    async def call(self, call: ToolCall) -> ToolOutput[JsonableCovariantT]:
-        """Call the tool using an LLM-provided ToolCall."""
-        raise NotImplementedError()
