+++ python/mirascope/llm/tools/tool.py
@@ -2,12 +2,19 @@
 
 from __future__ import annotations
 
-from collections.abc import Callable
+from collections.abc import Awaitable, Callable
 from dataclasses import dataclass
+from typing import TypeVar
 
 from ..content import ToolCall, ToolOutput
-from ..types import JsonableCovariantT, P
+from ..types import Jsonable, JsonableCovariantT, P
 from .base_tool import BaseTool
+
+ToolT = TypeVar(
+    "ToolT",
+    bound="Tool[..., Jsonable] | AsyncTool[..., Jsonable]",
+    covariant=True,
+)
 
 
 @dataclass
@@ -24,5 +31,23 @@ class Tool(BaseTool[P, JsonableCovariantT]):
     """The function that implements the tool's functionality."""
 
     def call(self, call: ToolCall) -> ToolOutput[JsonableCovariantT]:
+        """Call the tool using an LLM-provided ToolCall."""
+        raise NotImplementedError()
+
+
+@dataclass
+class AsyncTool(BaseTool[P, JsonableCovariantT]):
+    """Protocol defining an async tool that can be used by LLMs.
+
+    An AsyncTool represents an async function that can be called by an LLM during a call.
+    It includes metadata like name, description, and parameter schema.
+
+    This class is not instantiated directly but created by the `@tool()` decorator.
+    """
+
+    fn: Callable[P, Awaitable[JsonableCovariantT]]
+    """The async function that implements the tool's functionality."""
+
+    async def call(self, call: ToolCall) -> ToolOutput[JsonableCovariantT]:
         """Call the tool using an LLM-provided ToolCall."""
         raise NotImplementedError()
